trivial module renaming
