+++ themes/microfolio/source/js/DoubleBuffer.js
